% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{as_bilat_surface_vector}
\alias{as_bilat_surface_vector}
\title{Extract Bilateral Surface Data as a neurosurf Object}
\usage{
as_bilat_surface_vector(x, left_geometry, right_geometry)
}
\arguments{
\item{x}{A \code{\linkS4class{CiftiImage}} with loaded data.}

\item{left_geometry}{A \code{neurosurf::SurfaceGeometry} for the left hemisphere.}

\item{right_geometry}{A \code{neurosurf::SurfaceGeometry} for the right hemisphere.}
}
\value{
A \code{BilatNeuroSurfaceVector}.
}
\description{
Extracts left and right cortex data and combines them into a
\code{neurosurf::BilatNeuroSurfaceVector}.
}
