% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{as_neuro_vec}
\alias{as_neuro_vec}
\title{Extract All Subcortical Volume Data as a neuroim2 Object}
\usage{
as_neuro_vec(x)
}
\arguments{
\item{x}{A \code{\linkS4class{CiftiImage}} with loaded data.}
}
\value{
A \code{SparseNeuroVol} (single map) or \code{SparseNeuroVec}
  (multiple maps / time series).
}
\description{
Combines all volume-based brain structures into a single
\code{SparseNeuroVol} or \code{SparseNeuroVec}.
}
