% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_axis.R
\name{axes_to_header}
\alias{axes_to_header}
\title{Convert Axis Objects Back to CiftiHeader}
\usage{
axes_to_header(row_axis, col_axis, version = "2")
}
\arguments{
\item{row_axis}{A \code{\linkS4class{CiftiAxis}} for rows (dimension 0).}

\item{col_axis}{A \code{\linkS4class{CiftiAxis}} for columns (dimension 1).}

\item{version}{Character version string (default "2").}
}
\value{
A \code{\linkS4class{CiftiHeader}} object.
}
\description{
Convert Axis Objects Back to CiftiHeader
}
