% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_axis.R
\name{brain_model_from_mask}
\alias{brain_model_from_mask}
\title{Create a BrainModelAxis from a Mask}
\usage{
brain_model_from_mask(mask, name, affine = NULL)
}
\arguments{
\item{mask}{For surfaces: a logical vector (length = nvertex).
For volumes: a 3D logical array.}

\item{name}{Canonical CIFTI structure string.}

\item{affine}{4x4 affine matrix (required for volume masks).}
}
\value{
A \code{\linkS4class{BrainModelAxis}} object.
}
\description{
Create a BrainModelAxis from a Mask
}
