% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_axis.R
\name{brain_model_from_surface}
\alias{brain_model_from_surface}
\title{Create a BrainModelAxis from Surface Vertex Indices}
\usage{
brain_model_from_surface(vertices, nvertex, name)
}
\arguments{
\item{vertices}{Integer vector of 0-based vertex indices.}

\item{nvertex}{Total number of vertices on the surface.}

\item{name}{Canonical CIFTI structure string.}
}
\value{
A \code{\linkS4class{BrainModelAxis}} object.
}
\description{
Create a BrainModelAxis from Surface Vertex Indices
}
