% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{brain_models}
\alias{brain_models}
\title{Get Brain Models from an Object}
\usage{
brain_models(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A list of \code{\linkS4class{CiftiBrainModel}} objects.
}
\description{
Get Brain Models from an Object
}
