% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_xml.R
\name{build_cifti_xml}
\alias{build_cifti_xml}
\title{Build CIFTI-2 XML from Header}
\usage{
build_cifti_xml(header)
}
\arguments{
\item{header}{A \code{\linkS4class{CiftiHeader}} object.}
}
\value{
A character string containing the CIFTI-2 XML.
}
\description{
Converts a \code{\linkS4class{CiftiHeader}} to an XML character string.
}
