% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{cifti_data}
\alias{cifti_data}
\title{Get the Data Matrix}
\usage{
cifti_data(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A numeric matrix.
}
\description{
Get the Data Matrix
}
