% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{cifti_header}
\alias{cifti_header}
\title{Get the CIFTI Header}
\usage{
cifti_header(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A \code{\linkS4class{CiftiHeader}} object.
}
\description{
Get the CIFTI Header
}
