% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CIFTI_INDEX_TYPES}
\alias{CIFTI_INDEX_TYPES}
\title{CIFTI-2 Index Types}
\format{
An object of class \code{character} of length 5.
}
\usage{
CIFTI_INDEX_TYPES
}
\description{
Identifies what kind of data a dimension contains.
}
\keyword{datasets}
