% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{cifti_intent}
\alias{cifti_intent}
\title{Get the NIfTI Intent Code}
\usage{
cifti_intent(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
An integer intent code.
}
\description{
Get the NIfTI Intent Code
}
