% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CIFTI_INTENTS}
\alias{CIFTI_INTENTS}
\title{CIFTI-2 NIfTI Intent Codes}
\format{
An object of class \code{integer} of length 11.
}
\usage{
CIFTI_INTENTS
}
\description{
NIfTI intent codes used by CIFTI-2 files.
}
\keyword{datasets}
