% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CIFTI_MODEL_TYPES}
\alias{CIFTI_MODEL_TYPES}
\title{CIFTI-2 Brain Model Types}
\format{
An object of class \code{character} of length 2.
}
\usage{
CIFTI_MODEL_TYPES
}
\description{
CIFTI-2 Brain Model Types
}
\keyword{datasets}
