% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CIFTI_SERIES_UNITS}
\alias{CIFTI_SERIES_UNITS}
\title{CIFTI-2 Series Unit Types}
\format{
An object of class \code{character} of length 4.
}
\usage{
CIFTI_SERIES_UNITS
}
\description{
CIFTI-2 Series Unit Types
}
\keyword{datasets}
