% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CIFTI_STRUCTURES}
\alias{CIFTI_STRUCTURES}
\title{CIFTI-2 Brain Structure Names}
\format{
An object of class \code{character} of length 33.
}
\usage{
CIFTI_STRUCTURES
}
\description{
Named character vector of valid CIFTI-2 brain structure identifiers.
Names are short aliases; values are the full CIFTI XML strings.
}
\keyword{datasets}
