% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiHeader-class}
\alias{CiftiHeader-class}
\title{CiftiHeader Class}
\description{
Top-level parsed CIFTI-2 XML header.
}
\section{Slots}{

\describe{
\item{\code{version}}{Character string (e.g., "2").}

\item{\code{matrix}}{A \code{\linkS4class{CiftiMatrix}} object.}
}}

