% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiLabelTable-class}
\alias{CiftiLabelTable-class}
\title{CiftiLabelTable Class}
\description{
A table of label entries for a named map.
}
\section{Slots}{

\describe{
\item{\code{labels}}{A list of \code{\linkS4class{CiftiLabel}} objects.}
}}

