% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiMatrix-class}
\alias{CiftiMatrix-class}
\title{CiftiMatrix Class}
\description{
Complete matrix metadata — contains all dimension descriptors.
}
\section{Slots}{

\describe{
\item{\code{maps}}{A list of \code{\linkS4class{CiftiMatrixIndicesMap}} objects.}

\item{\code{metadata}}{A named list of metadata key-value pairs.}
}}

