% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiMatrixIndicesMap-class}
\alias{CiftiMatrixIndicesMap-class}
\title{CiftiMatrixIndicesMap Class}
\description{
Describes one dimension of the CIFTI data matrix.
}
\section{Slots}{

\describe{
\item{\code{applies_to_dimension}}{Integer vector (0-based dimension indices this map applies to).}

\item{\code{index_type}}{Character: one of the CIFTI_INDEX_TYPES values.}

\item{\code{brain_models}}{List of \code{\linkS4class{CiftiBrainModel}} objects.}

\item{\code{named_maps}}{List of \code{\linkS4class{CiftiNamedMap}} objects.}

\item{\code{parcels}}{List of \code{\linkS4class{CiftiParcel}} objects.}

\item{\code{surfaces}}{List of \code{\linkS4class{CiftiSurface}} objects.}

\item{\code{volume}}{A \code{\linkS4class{CiftiVolume}} or NULL.}

\item{\code{series_start}}{Numeric start value for SERIES type.}

\item{\code{series_step}}{Numeric step value for SERIES type.}

\item{\code{series_unit}}{Character unit for SERIES type.}

\item{\code{number_of_series_points}}{Integer count for SERIES type.}
}}

