% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiNamedMap-class}
\alias{CiftiNamedMap-class}
\title{CiftiNamedMap Class}
\description{
One named map entry (used in SCALARS and LABELS index types).
}
\section{Slots}{

\describe{
\item{\code{map_name}}{Character string naming the map.}

\item{\code{label_table}}{A \code{\linkS4class{CiftiLabelTable}} or NULL.}

\item{\code{metadata}}{A named list of metadata key-value pairs.}
}}

