% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiParcel-class}
\alias{CiftiParcel-class}
\title{CiftiParcel Class}
\description{
One parcel in a PARCELS index map.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character parcel name.}

\item{\code{surface_vertices}}{A named list mapping structure names to integer vectors
of 0-based vertex indices.}

\item{\code{voxel_indices_ijk}}{Integer matrix (N x 3) of 0-based IJK voxel indices.}
}}

