% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiSurface-class}
\alias{CiftiSurface-class}
\title{CiftiSurface Class}
\description{
Surface metadata for a PARCELS index map.
}
\section{Slots}{

\describe{
\item{\code{brain_structure}}{Character: canonical CIFTI structure string.}

\item{\code{number_of_vertices}}{Integer: total vertices on the surface.}
}}

