% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R
\docType{class}
\name{CiftiVolume-class}
\alias{CiftiVolume-class}
\title{CiftiVolume Class}
\description{
Volume geometry reference (dimensions and affine transform).
}
\section{Slots}{

\describe{
\item{\code{dimensions}}{Integer vector of length 3 (I, J, K).}

\item{\code{transform}}{A 4x4 numeric matrix (voxel-to-world affine).}
}}

