% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{col_axis}
\alias{col_axis}
\title{Get the Column Axis of a CIFTI Object}
\usage{
col_axis(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A \code{\linkS4class{CiftiAxis}} object.
}
\description{
Get the Column Axis of a CIFTI Object
}
