% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{dconn}
\alias{dconn}
\title{Create a Dense Connectivity CIFTI Image}
\usage{
dconn(data, row_brain_model, col_brain_model = row_brain_model)
}
\arguments{
\item{data}{Numeric matrix (grayordinates x grayordinates).}

\item{row_brain_model}{A \code{\linkS4class{BrainModelAxis}} for rows.}

\item{col_brain_model}{A \code{\linkS4class{BrainModelAxis}} for columns
(defaults to \code{row_brain_model}).}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3001.
}
\description{
Connectivity matrix between grayordinates (BrainModelAxis x BrainModelAxis).
}
