% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{dlabel}
\alias{dlabel}
\title{Create a Dense Label CIFTI Image}
\usage{
dlabel(data, brain_model, names = NULL, label_tables = NULL)
}
\arguments{
\item{data}{Integer matrix (maps x grayordinates) of label keys.}

\item{brain_model}{A \code{\linkS4class{BrainModelAxis}} for columns.}

\item{names}{Character vector of map names.}

\item{label_tables}{List of \code{\linkS4class{CiftiLabelTable}} objects.}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3007.
}
\description{
Create a Dense Label CIFTI Image
}
