% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_axis.R
\name{.axis_to_mim}
\alias{.axis_to_mim}
\title{Convert an Axis back to a MatrixIndicesMap}
\usage{
.axis_to_mim(axis, dimension)
}
\description{
Convert an Axis back to a MatrixIndicesMap
}
\keyword{internal}
