% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_xml.R
\name{.build_matrix_indices_map}
\alias{.build_matrix_indices_map}
\title{Build <MatrixIndicesMap> XML node}
\usage{
.build_matrix_indices_map(parent, mim)
}
\description{
Build <MatrixIndicesMap> XML node
}
\keyword{internal}
