% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_io.R
\name{.read_cifti_data}
\alias{.read_cifti_data}
\title{Read binary data from a CIFTI file}
\usage{
.read_cifti_data(file_name, vox_offset, datatype, bitpix, nrows, ncols)
}
\description{
Read binary data from a CIFTI file
}
\keyword{internal}
