% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_io.R
\name{.read_cifti_extension}
\alias{.read_cifti_extension}
\title{Read the CIFTI XML extension from a NIfTI-2 file}
\usage{
.read_cifti_extension(file_name, vox_offset)
}
\description{
Read the CIFTI XML extension from a NIfTI-2 file
}
\keyword{internal}
