% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_io.R
\name{.write_int64}
\alias{.write_int64}
\title{Write an int64 value as 8 raw bytes (little-endian)}
\usage{
.write_int64(con, value)
}
\description{
Write an int64 value as 8 raw bytes (little-endian)
}
\keyword{internal}
