% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_io.R
\name{.write_nifti2_header}
\alias{.write_nifti2_header}
\title{Write a NIfTI-2 header (540 bytes, exact layout per spec)}
\usage{
.write_nifti2_header(con, intent, datatype, bitpix, nrows, ncols, vox_offset)
}
\description{
Write a NIfTI-2 header (540 bytes, exact layout per spec)
}
\keyword{internal}
