% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{dpconn}
\alias{dpconn}
\title{Create a Dense-to-Parcel Connectivity CIFTI Image}
\usage{
dpconn(data, brain_model, parcels)
}
\arguments{
\item{data}{Numeric matrix (grayordinates x parcels).}

\item{brain_model}{A \code{\linkS4class{BrainModelAxis}} for rows.}

\item{parcels}{A \code{\linkS4class{ParcelsAxis}} for columns.}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3010.
}
\description{
Connectivity from grayordinates (rows) to parcels (columns).
}
