% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{dscalar}
\alias{dscalar}
\title{Create a Dense Scalar CIFTI Image}
\usage{
dscalar(data, brain_model, names = NULL)
}
\arguments{
\item{data}{Numeric matrix (maps x grayordinates).}

\item{brain_model}{A \code{\linkS4class{BrainModelAxis}} for columns.}

\item{names}{Character vector of map names.}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3006.
}
\description{
Create a Dense Scalar CIFTI Image
}
