% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{dtseries}
\alias{dtseries}
\title{Create a Dense Timeseries CIFTI Image}
\usage{
dtseries(data, brain_model, start = 0, step = 1, unit = "SECOND")
}
\arguments{
\item{data}{Numeric matrix (time x grayordinates).}

\item{brain_model}{A \code{\linkS4class{BrainModelAxis}} for columns.}

\item{start}{Numeric start time (default 0).}

\item{step}{Numeric TR / time step.}

\item{unit}{Character unit (default "SECOND").}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3002.
}
\description{
Create a Dense Timeseries CIFTI Image
}
