% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_axis.R
\name{header_to_axes}
\alias{header_to_axes}
\title{Convert CiftiHeader to Axis Objects}
\usage{
header_to_axes(header)
}
\arguments{
\item{header}{A \code{\linkS4class{CiftiHeader}} object.}
}
\value{
A list with elements \code{row_axis} and \code{col_axis}.
}
\description{
Convert CiftiHeader to Axis Objects
}
