% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ijk_to_linear}
\alias{ijk_to_linear}
\title{IJK Voxel Indices to Linear Index}
\usage{
ijk_to_linear(ijk, vol_dims)
}
\arguments{
\item{ijk}{An integer matrix (N x 3) of 0-based IJK coordinates.}

\item{vol_dims}{An integer vector of length 3 (volume dimensions).}
}
\value{
An integer vector of 1-based linear indices.
}
\description{
Converts 0-based IJK voxel coordinates to a 1-based linear index
given volume dimensions.
}
\keyword{internal}
