% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_surface_structure}
\alias{is_surface_structure}
\title{Check if a Structure is a Surface Type}
\usage{
is_surface_structure(structure)
}
\arguments{
\item{structure}{A canonical CIFTI structure string.}
}
\value{
Logical.
}
\description{
Check if a Structure is a Surface Type
}
\keyword{internal}
