% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{iter_structures}
\alias{iter_structures}
\title{Iterate Over Brain Structures}
\usage{
iter_structures(x, ...)
}
\arguments{
\item{x}{An axis object.}

\item{...}{Additional arguments.}
}
\value{
A named list of index ranges.
}
\description{
Returns a list of (structure_name, slice) pairs for each brain structure
in an axis.
}
