% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linear_to_ijk}
\alias{linear_to_ijk}
\title{Linear Index to IJK Voxel Indices}
\usage{
linear_to_ijk(idx, vol_dims)
}
\arguments{
\item{idx}{An integer vector of 1-based linear indices.}

\item{vol_dims}{An integer vector of length 3.}
}
\value{
An integer matrix (N x 3) of 0-based IJK coordinates.
}
\description{
Converts 1-based linear index to 0-based IJK voxel coordinates.
}
\keyword{internal}
