% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{named_maps}
\alias{named_maps}
\title{Get Named Maps from an Object}
\usage{
named_maps(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A list of \code{\linkS4class{CiftiNamedMap}} objects.
}
\description{
Get Named Maps from an Object
}
