% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NIFTI2_DTYPE_SIZES}
\alias{NIFTI2_DTYPE_SIZES}
\title{Size in bytes for NIfTI-2 data types}
\format{
An object of class \code{integer} of length 5.
}
\usage{
NIFTI2_DTYPE_SIZES
}
\description{
Size in bytes for NIfTI-2 data types
}
\keyword{internal}
