% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NIFTI2_DTYPES}
\alias{NIFTI2_DTYPES}
\title{NIfTI-2 Data Type Codes}
\format{
An object of class \code{integer} of length 5.
}
\usage{
NIFTI2_DTYPES
}
\description{
Numeric type codes for NIfTI-2 binary data.
}
\keyword{internal}
