% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NIFTI2_HEADER_SIZE}
\alias{NIFTI2_HEADER_SIZE}
\title{NIfTI-2 header size in bytes}
\format{
An object of class \code{integer} of length 1.
}
\usage{
NIFTI2_HEADER_SIZE
}
\description{
NIfTI-2 header size in bytes
}
\keyword{internal}
