% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_to_16}
\alias{pad_to_16}
\title{Pad Bytes to 16-byte Boundary}
\usage{
pad_to_16(n)
}
\arguments{
\item{n}{Number of bytes.}
}
\value{
The next multiple of 16 >= n.
}
\description{
Pad Bytes to 16-byte Boundary
}
\keyword{internal}
