% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_axis.R
\name{parcels_from_brain_models}
\alias{parcels_from_brain_models}
\title{Create a ParcelsAxis from Brain Models}
\usage{
parcels_from_brain_models(named_brain_models)
}
\arguments{
\item{named_brain_models}{A named list where each element is a
\code{\linkS4class{BrainModelAxis}} and names are parcel names.}
}
\value{
A \code{\linkS4class{ParcelsAxis}} object.
}
\description{
Create a ParcelsAxis from Brain Models
}
