% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_xml.R
\name{parse_cifti_xml}
\alias{parse_cifti_xml}
\title{Parse CIFTI-2 XML Header}
\usage{
parse_cifti_xml(xml_data)
}
\arguments{
\item{xml_data}{Raw vector, character string, or xml2 document.}
}
\value{
A \code{\linkS4class{CiftiHeader}} object.
}
\description{
Parses raw XML bytes (or a character string) from a CIFTI-2 NIfTI extension
into a \code{\linkS4class{CiftiHeader}} object.
}
