% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{pconn}
\alias{pconn}
\title{Create a Parcel Connectivity CIFTI Image}
\usage{
pconn(data, row_parcels, col_parcels = row_parcels)
}
\arguments{
\item{data}{Numeric matrix (parcels x parcels).}

\item{row_parcels}{A \code{\linkS4class{ParcelsAxis}} for rows.}

\item{col_parcels}{A \code{\linkS4class{ParcelsAxis}} for columns
(defaults to \code{row_parcels}).}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3003.
}
\description{
Connectivity matrix between parcels (ParcelsAxis x ParcelsAxis).
}
