% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{pdconn}
\alias{pdconn}
\title{Create a Parcel-to-Dense Connectivity CIFTI Image}
\usage{
pdconn(data, parcels, brain_model)
}
\arguments{
\item{data}{Numeric matrix (parcels x grayordinates).}

\item{parcels}{A \code{\linkS4class{ParcelsAxis}} for rows.}

\item{brain_model}{A \code{\linkS4class{BrainModelAxis}} for columns.}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3011.
}
\description{
Connectivity from parcels (rows) to grayordinates (columns).
}
