% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{plabel}
\alias{plabel}
\title{Create a Parcel Label CIFTI Image}
\usage{
plabel(data, parcels, names = NULL, label_tables = NULL)
}
\arguments{
\item{data}{Integer matrix (maps x parcels) of label keys.}

\item{parcels}{A \code{\linkS4class{ParcelsAxis}} for columns.}

\item{names}{Character vector of map names.}

\item{label_tables}{List of \code{\linkS4class{CiftiLabelTable}} objects.}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3012.
}
\description{
Create a Parcel Label CIFTI Image
}
