% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{pscalar}
\alias{pscalar}
\title{Create a Parcel Scalar CIFTI Image}
\usage{
pscalar(data, parcels, names = NULL)
}
\arguments{
\item{data}{Numeric matrix (maps x parcels).}

\item{parcels}{A \code{\linkS4class{ParcelsAxis}} for columns.}

\item{names}{Character vector of map names.}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3008.
}
\description{
Create a Parcel Scalar CIFTI Image
}
