% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_intent.R
\name{ptseries}
\alias{ptseries}
\title{Create a Parcel Timeseries CIFTI Image}
\usage{
ptseries(data, parcels, start = 0, step = 1, unit = "SECOND")
}
\arguments{
\item{data}{Numeric matrix (time x parcels).}

\item{parcels}{A \code{\linkS4class{ParcelsAxis}} for columns.}

\item{start}{Numeric start time (default 0).}

\item{step}{Numeric TR / time step.}

\item{unit}{Character unit (default "SECOND").}
}
\value{
A \code{\linkS4class{CiftiImage}} with intent 3004.
}
\description{
Create a Parcel Timeseries CIFTI Image
}
