% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{row_axis}
\alias{row_axis}
\title{Get the Row Axis of a CIFTI Object}
\usage{
row_axis(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A \code{\linkS4class{CiftiAxis}} object.
}
\description{
Get the Row Axis of a CIFTI Object
}
