% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_class.R, R/cifti_axis_methods.R
\docType{class}
\name{SeriesAxis-class}
\alias{SeriesAxis-class}
\alias{show,SeriesAxis-method}
\alias{length,SeriesAxis-method}
\alias{[,SeriesAxis,numeric,missing-method}
\alias{c,SeriesAxis-method}
\alias{==,SeriesAxis,SeriesAxis-method}
\title{SeriesAxis Class}
\usage{
\S4method{show}{SeriesAxis}(object)

\S4method{length}{SeriesAxis}(x)

\S4method{[}{SeriesAxis,numeric,missing}(x, i, j, ..., drop = TRUE)

\S4method{c}{SeriesAxis}(x, ...)

\S4method{==}{SeriesAxis,SeriesAxis}(e1, e2)
}
\arguments{
\item{object}{A \code{SeriesAxis} object.}

\item{x}{A \code{SeriesAxis} object.}

\item{i}{Numeric index vector.}

\item{j}{Not used.}

\item{...}{Additional arguments (ignored).}

\item{drop}{Logical (ignored).}

\item{e1}{First \code{SeriesAxis}.}

\item{e2}{Second \code{SeriesAxis}.}
}
\description{
Describes a SERIES dimension (e.g., time points).
}
\section{Slots}{

\describe{
\item{\code{start}}{Numeric start value.}

\item{\code{step}}{Numeric step size.}

\item{\code{size}}{Integer number of points.}

\item{\code{unit}}{Character unit string (one of CIFTI_SERIES_UNITS).}
}}

