% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_cifti_index}
\alias{to_cifti_index}
\title{Convert 1-based R Index to 0-based CIFTI Index}
\usage{
to_cifti_index(x)
}
\arguments{
\item{x}{An integer vector of 1-based indices.}
}
\value{
An integer vector of 0-based indices.
}
\description{
Convert 1-based R Index to 0-based CIFTI Index
}
