% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_r_index}
\alias{to_r_index}
\title{Convert 0-based CIFTI Index to 1-based R Index}
\usage{
to_r_index(x)
}
\arguments{
\item{x}{An integer vector of 0-based indices.}
}
\value{
An integer vector of 1-based indices.
}
\description{
Convert 0-based CIFTI Index to 1-based R Index
}
