% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{volume_info}
\alias{volume_info}
\title{Get Volume Information}
\usage{
volume_info(x, ...)
}
\arguments{
\item{x}{A CIFTI object.}

\item{...}{Additional arguments.}
}
\value{
A \code{\linkS4class{CiftiVolume}} object or NULL.
}
\description{
Get Volume Information
}
