% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cifti_io.R
\name{write_cifti}
\alias{write_cifti}
\title{Write a CIFTI-2 File}
\usage{
write_cifti(x, file_name, data_type = "FLOAT32")
}
\arguments{
\item{x}{A \code{\linkS4class{CiftiImage}} object.}

\item{file_name}{Output file path.}

\item{data_type}{Data type for the binary data. One of "FLOAT32" (default)
or "FLOAT64".}
}
\value{
Invisibly returns the file path.
}
\description{
Writes a \code{\linkS4class{CiftiImage}} to a CIFTI-2 NIfTI file.
}
